+++ packages/deep-heating-rx/src/lib/streams/trvs/trvTemperatures.ts
@@ -3,8 +3,11 @@ import {
   TrvUpdate,
 } from '@home-automation/deep-heating-types';
 import { shareReplayLatestDistinctByKey } from '@home-automation/rxx';
+import debug from 'debug';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, tap } from 'rxjs/operators';
+
+const log = debug('deep-heating:trv-temp-flow');
 
 export const getTrvTemperatures = (
   trvUpdates: Observable<TrvUpdate>,
@@ -14,5 +17,12 @@ export const getTrvTemperatures = (
       climateEntityId: x.climateEntityId,
       temperatureReading: x.state.temperature,
     })),
+    tap((x) => {
+      log(
+        '[1-trvTemperatures] %s: %d°C',
+        x.climateEntityId,
+        x.temperatureReading.temperature,
+      );
+    }),
     shareReplayLatestDistinctByKey((x) => x.climateEntityId),
   );