+++ packages/deep-heating-web/src/lib/components/Spinner.svelte
@@ -1,1 +1,1 @@
-<button class="btn btn-circle loading" />
+<button class="btn btn-circle loading" aria-label="Loading"></button>