+++ packages/deep-heating-web/src/routes/+layout.svelte
@@ -4,14 +4,20 @@
   import { homeStore } from '$lib/stores/home';
   import '../app.css';
   import type { LayoutData } from './$types';
+  import type { Snippet } from 'svelte';
 
-  export let data: LayoutData;
+  interface Props {
+    data: LayoutData;
+    children: Snippet;
+  }
+
+  let { data, children }: Props = $props();
 
   $appSettingsStore = data;
 </script>
 
 <div class="container mx-auto">
-  <slot />
+  {@render children()}
 
   {#if !$homeStore.connected}
     <div