+++ packages/deep-heating-web/svelte.config.js
@@ -1,4 +1,4 @@
-import adapter from '@sveltejs/adapter-node';
+import adapter from 'svelte-adapter-bun';
 import * as path from 'path';
 import preprocess from 'svelte-preprocess';